separated unit tests into three separate files
